+++ .gitignore
@@ -3,3 +3,4 @@
 .idea/MoviesLand-ECF-DWWM.iml
 .idea/vcs.xml
 .idea/workspace.xml
+*.xml
