+++ src/Views/addMovie.php
@@ -1,4 +1,4 @@
-<?php include(__DIR__ . "./../template/header.php");
+<?php include(__DIR__ . "./../Templates/header.php");
 
 ?>
 
@@ -45,4 +45,4 @@
 </form>
 
 
-<?php include(__DIR__ . "./../template/footer.php"); ?>
+<?php include(__DIR__ . "./../Templates/footer.php"); ?>
