+++ src/Views/home.php
@@ -1,8 +1,8 @@
-<?php include (__DIR__ . "./../template/header.php"); ?>
+<?php include (__DIR__ . "./../Templates/header.php"); ?>
 
 <div><img src="https://i.ytimg.com/vi/iJNbQtItYFo/maxresdefault.jpg" alt="I Love movies but i hate the cinema" width="80%"
     height="auto"></div>
 
 
 
-<?php include (__DIR__ . "./../template/footer.php"); ?>
+<?php include (__DIR__ . "./../Templates/footer.php"); ?>
