+++ src/Views/movie.php
@@ -1,4 +1,4 @@
-<?php include(__DIR__ . "./../template/header.php"); ?>
+<?php include(__DIR__ . "./../Templates/header.php"); ?>
 
 
 <div class="row row-cols-1 row-cols-md-4">
@@ -17,4 +17,4 @@
 <div></div>
 
 
-<?php include(__DIR__ . "./../template/footer.php"); ?>
+<?php include(__DIR__ . "./../Templates/footer.php"); ?>
